Enemy improvement

+++ src/main/java/org/example/Enemy1.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class Enemy1 extends Enemy{
     int id;
-    public Enemy1(double speed, int totHP, double x, double y,int id){
+    public Enemy1(double speed, int totHP, int x, int y,int id){
         super(speed,totHP,x,y);
         this.id =id;
     }
